backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
Pillow
SQLAlchemy
PyMySQL
